almost last three weeks ago successful lmported product.
Should consider before sending that “I quit!” email that you’ve been fantasizing about:

Identify a problem that you feel driven to solve.

Starting a business is not easy, and scaling it is even harder. But the strongest fuel is a personal connection to what you’re doing. It could be that you have experience working in an industry and understand its shortcomings firsthand. Or perhaps you’re part of a consumer segment that’s underserved by the current offerings. Maybe you’re simply met with a very specific frustration every day, that others are sure to share. However you come to your idea, you should feel like you have no choice but to start this particular business at this moment in time. It will make the mornings when you wake up and wish that it was someone else’s problem much easier to bear.

Consider your role as founder.

More than ever, people care deeply about who’s behind the companies they’re purchasing from. It’s hard to feel a personal connection to a nameless, faceless corporation, and far more rewarding to support brands that are built by individuals with a compelling story. Particularly on social media, so many brands gain traction by having their founders front and center as part of the narrative: speaking to their experiences, demonstrating humility and vulnerability, and putting a human face to the business. This doesn’t mean that in order to start a company you need to be prepared to be a public persona who reveals every aspect of your private life.to get your own startup :calls or whatsApps us 08037938828. Serikijoshua4@gmail.com